+++ packages/open-lens/package.json
@@ -206,7 +206,7 @@
     "@k8slens/messaging": "^1.0.0-alpha.1",
     "@k8slens/messaging-for-main": "^1.0.0-alpha.1",
     "@k8slens/messaging-for-renderer": "^1.0.0-alpha.1",
-    "@k8slens/react-application-root": "^1.0.0-alpha.0",
+    "@k8slens/react-application": "^1.0.0-alpha.0",
     "@k8slens/run-many": "^1.0.0-alpha.1",
     "@k8slens/startable-stoppable": "^1.0.0-alpha.1",
     "@k8slens/test-utils": "^1.0.0-alpha.1",
