+++ packages/technical-features/react-application/.eslintrc.json
