+++ packages/technical-features/react-application/.prettierrc
