+++ packages/technical-features/react-application/jest.config.js
