+++ packages/technical-features/react-application/src/react-application/react-application-higher-order-component-injection-token.ts
