+++ packages/technical-features/react-application/src/react-application/react-application.tsx
