+++ packages/technical-features/react-application/src/render-application/render-application-when-application-is-ready.injectable.tsx
