+++ packages/technical-features/react-application/src/render-application/render.injectable.tsx
