+++ packages/technical-features/react-application/tsconfig.json
