+++ packages/technical-features/react-application/webpack.config.js
